fix(expect): ensure final LAVA line is captured
Some LAVA expectations include lines for which no newline is printed,
and for some reason LAVA will not capture the final line if it doesn't
end in a newline.

This change ensures that these LAVA fragments explicitly look for a
newline before returning a successful result. This ensures that their
matching Expect scripts pass in the OpenCI.

Signed-off-by: Chris Kay <chris.kay@arm.com>
Change-Id: I7dcce858f809c58112695e3262cbc4ee915d2289