refactor: use updated LAVA log names in post-LAVA Expect scripts
The log splitting script (part of the CI jobs repository) now directs
LAVA terminal output to files of the form `lava-${NAME}.log`, where
`NAME` is the name of the terminal captured by the `NAME` regular
expression groups specified by the feedbacks in the LAVA job definition.

A recent edge-case showed that platforms with duplicate terminal names
may exist, which means we must now make several adjustments to how we
enumerate the terminals and name our feedbacks.

The most prominent change is that feedback terminals in the LAVA job
definitions now use the port as the feedback name. The port is the only
mechanism we have for uniquely identifying a UART so given its FVP
terminal printout so, while it isn't pretty, it is at least resilient.

Signed-off-by: Chris Kay <chris.kay@arm.com>
Change-Id: Ib6222e6342e7d795ea1f3f603ba1fed9405cb81b