refactor: move some script variables to `openci-env.sh`
There's not much consistency to how the three CIs are configured, and
we're encountering an issue with the post-LAVA Expect scripts where
`nfs_volume` isn't set when sourcing `utils.sh`. This change moves three
frequently-used variablles - `nfs_volume`, `jenkins_url`, and
`tfa_downloads` into an OpenCI environment script as is done for the
internal CI.

It's not a huge clean-up, but it's a decent start that we need for the
post-LAVA Expect scripts.

Signed-off-by: Chris Kay <chris.kay@arm.com>
Change-Id: I22f3f806c4effc0e4ea529da8a9532d70b44b0e0