ci(rng-trap): removed expect and modified run config
Since the trap triggered by RNDR and RNDRRS reads is not unhandled
anymore, the run config was changed in order not to expect a panic
but to allow the test to finish returning test result value.
In addition, FEAT_RNG option has been added to the model so it can
now process RNDR and RNDRRS reads from EL3 and therefore random
numbers can be obtained upon those reads.

Signed-off-by: Juan Pablo Conde <juanpablo.conde@arm.com>
Change-Id: I6d78889c45af73391359d09a34c52bd013c7316d