scp-boot-tests: Remove TC0 configuration
A recent change in TF-A [8c276af3902a5312e9ef44a3316cce2caaad7c64]
has caused the TC0 linux boot to fail. In the meantime,
support for TC0 has been dropped (deprecated).
As a result, we are forced to remove the boot test for TC0 given
that it has reached the end of its support.

Signed-off-by: Leandro Belli <leandro.belli@arm.com>
Change-Id: I782e65839460f56b5d6339edbb8362efee200ee0